Menu Setup
The menu UI is in theory working properly in 2D. We need to switch to physical buttons instead of 2D buttons. Right now, it automatically connects you to a room (Room 1), instead of typing out which room name you want with displays of all current active rooms. Hotte scene hehe