Lever Works, UI Canvas, Billboard
Levers should in theory work and it doesn't keep adding more players that we don't have. Added a UI canvas in the locker room scene that displays to everyone the amount of players ready and needed to start. Made a Billboard script so that the UI always follows the player no matter what.