Post-Game Leaderboards
Added the end-of-round leaderboards that appear at the end of the round. The leaderboard is not sorted by highest kills as of right now. I created a reference to the last scene in the Game Manager. If the last scene was the arena scene (just came back from a game), then we open up the leaderboards.